requirements.txt
django==3.2.5
djangorestframework==3.12.4
drf-nested-routers==0.93.3
drf-jwt==1.19.1
psycopg2==2.9.1
django-filter==2.4.0
